+++ backend/controller/post.js
@@ -23,24 +23,31 @@ export const addPost = (req, res) => {
   const token = req.cookies.accessToken;
   if (!token) return res.status(401).json("Not Logged in");
 
-  jwt.verify(token, "secretkey", (err, userInfo) => { 
+  jwt.verify(token, "secretkey", (err, userInfo) => {
     if (err) return res.status(403).json("Token is not valid");
     if (!req.file) return res.status(400).json("No file uploaded.");
 
-    const imgPath = req.file.path;
-    const q =
-      "INSERT INTO posts (`desc`, `img`, `createdAt`, `userId`) VALUES (?, ?, ?, ?)";
-
-    const values = [
-      req.body.desc,
-      imgPath,
-      moment(Date.now()).format("YYYY-MM-DD HH:mm:ss"),
-      userInfo.id,
-    ];
-    db.query(q, values, (err, data) => {
-      // 使用参数化查询
+    const qCheckRevoked = "SELECT isRevoked FROM login WHERE id = ?";
+    db.query(qCheckRevoked, [userInfo.id], (err, data) => {
       if (err) return res.status(500).json(err);
-      return res.status(200).json("Post has been created");
+      if (data[0].isRevoked) return res.status(403).json("Account is revoked");
+
+      const imgPath = req.file.path;
+      const q =
+        "INSERT INTO posts (`desc`, `img`, `createdAt`, `userId`) VALUES (?, ?, ?, ?)";
+
+      const values = [
+        req.body.desc,
+        imgPath,
+        moment(Date.now()).format("YYYY-MM-DD HH:mm:ss"),
+        userInfo.id,
+      ];
+
+      db.query(q, values, (err, data) => {
+        // 使用参数化查询
+        if (err) return res.status(500).json(err);
+        return res.status(200).json("Post has been created");
+      });
     });
   });
 };
@@ -51,16 +58,21 @@ export const deletePost = (req, res) => {
 
   jwt.verify(token, "secretkey", (err, userInfo) => {
     if (err) return res.status(403).json("Token is not valid!");
-
-    const q = "DELETE FROM posts WHERE `id`=? AND `userId` = ?";
-
-    const values = [req.params.id, userInfo.id];
-
-    db.query(q, values, (err, data) => {
+    const qCheckRevoked = "SELECT isRevoked FROM login WHERE id = ?";
+    db.query(qCheckRevoked, [userInfo.id], (err, data) => {
       if (err) return res.status(500).json(err);
-      if (data.affectedRows > 0)
-        return res.status(200).json("Post has been deleted.");
-      return res.status(403).json("You can delete only your post");
+      if (data[0].isRevoked) return res.status(403).json("Account is revoked");
+
+      const q = "DELETE FROM posts WHERE `id`=? AND `userId` = ?";
+
+      const values = [req.params.id, userInfo.id];
+
+      db.query(q, values, (err, data) => {
+        if (err) return res.status(500).json(err);
+        if (data.affectedRows > 0)
+          return res.status(200).json("Post has been deleted.");
+        return res.status(403).json("You can delete only your post");
+      });
     });
   });
 };
